Product catalog and dialog now displaying items from database.
Realized I needed to switch to a persistent database, as switching
between dialogs resulted in loss of memory based databased changes.
There still, however, seems to be the need for a single database
connection that can be passed to the dialog windows when they need
access to display or edit the database.  Right now, using the default
QSqlDatabase::addDatabase() function, which results in disconnecting the
previous connection and establishing a new one each time.

 On branch devworking
 Your branch is up-to-date with 'origin/devworking'.

 Changes to be committed:
	modified:   catalogue.cpp
	modified:   databasedialog.cpp
	modified:   databasedialog.h
	modified:   mainwindow.cpp
	modified:   old-bar-pi.pro
	modified:   productdialog.cpp
	modified:   productdialog.h