Tweaked the database to use integer for the retired field in the inventory tabel.  Was boolean, but SQLITE doesn't support boolean as a type.  In conjunction with change to integer, created a 'yesno' table as a reference so that the retired field would display "Yes" or "No".
Tried to better control the state of the push buttons on the catalog
dialog. During this process, decided that a regular table view of
the database would be a better user interface.  Getting ready to
scrap the widget dialog and data mapper and go to a table view
like the inventory has.

 On branch inv

 Changes to be committed:
	modified:   catalog.cpp
	modified:   catalog.h
	modified:   databasedialog.cpp
	modified:   inventory.cpp
	modified:   inventory.h
	modified:   mainwindow.cpp