Added code to compute liquor product usage based on menu sales.
Added summary of menu cocktails.  For some (not all) of the cocktails, I add
their main liquor into the pour total for that product.  Need data for remaining
cocktails regarding their ingredients, to complete liquor product usage
calculation.

For some cocktails that have more than one ingredient that is being tracked,
a method needs to be added to add multiple items from a single entry to the
pour counts.