Separated Stock Manager into Inventory Manager and Invoice Manager. Modified stockmanager to be inventorymanager only.  Have not yet modified invoicemanager, nor added a tab in the GUI for it.
Changed some GUI buttons to be member variables so that they could
be enabled/disabled, according to the state of the process.  Now
when the Start button is clicked, it is disabled (can't reclick
after starting an inventory.)  There are separate buttons for
Commit, which writes inventory to the database, and Report, which
creates a CSV file.  Clicking Commit or Report disables that button,
so that one cannot write the inventory to the database more than once,
and there's no point in rewriting the CSV file.

Clicking Reset clears the tally table and enables the Start, Commit,
and Report buttons.

 On branch secondgen
 Changes to be committed:

	new file:   Database Design.ods
	modified:   barpi_app/CMakeLists.txt
	new file:   barpi_app/inventorymanager.cpp
	new file:   barpi_app/inventorymanager.h
	renamed:    barpi_app/stockmanager.cpp -> barpi_app/invoicemanager.cpp
	renamed:    barpi_app/stockmanager.h -> barpi_app/invoicemanager.h
	modified:   barpi_app/mainwindow.cpp
	modified:   barpi_app/mainwindow.h