Implemented a date delegate to properly display and edit the date field in the invoice table.
Still getting a database error when attaching to the invoice and
invoicedetails tables in the database.  Can't see that the code is
any different for the products table (catalog class).

Rethinking the use of the QSqlRelationalTableModel for the invoice
and invoicedetails tables.  The SQL table model doesn't offer signals
to notify when the selected cell changes, which is needed to map the
invoice to the invoicedetails.  A possible alternative is to use
QTableWidget, and populate the table with SQL queries.

 On branch secondgen
 Changes to be committed:

	modified:   ../Database Design.ods
	modified:   CMakeLists.txt
	modified:   databasedialog.cpp
	new file:   datedelegate.cpp
	new file:   datedelegate.h
	modified:   invoicemanager.cpp
	modified:   invoicemanager.h
	modified:   mainwindow.cpp